requirements.txt
torch>=2.0.0
transformers>=4.30.0
sentencepiece
protobuf
pymupdf
trafilatura
streamlit
pandas
plotly
pyyaml
emoji
